ajustes nos mapeamentos:   - comissão: acrescentado check de pago na tabela de comissão.   - Pedido: alterado o tamanho do campo de observação para 250 caracteres. Alterada o campo data_reg_pedido de int para datetime.   - ItemPedido: alterado o tamanho do campo de obsercao para 250 caracteres.

+++ KIS.System.Advanced.Domain/Entities/ItemPedido.cs
@@ -15,8 +15,8 @@ namespace KIS.System.Advanced.Domain.Entities
         public Produto Produto { get; set; }
         public string OBS_PRODUTO { get; set; }
         public int QTD_PEDIDO { get; set; }
-        public float VALOR_UN_PEDIDO { get; set; }
-        public float DESCONTO_PEDIDO { get; set; }
+        public Double VALOR_UN_PEDIDO { get; set; }
+        public Double DESCONTO_PEDIDO { get; set; }
 
     }
 }
